Feature/100 (#129)
* Initial restructure of models file into a module.

* Added __init__ file to make the directory a module.

* Removed register_model method and calls.

* Added dedicated constatants file. #104

* Updated import statements.

* Added missing import for numpy.

* Renamed CONSTANTS to constants. #100

* Merged EmbeddingModel with develop.

* Merged RandomBaseline with develop.

* Merged TransE with develop.

* Merged DistMult with develop.

* Merged ComplEx with develop.

* Merged HolE with develop.

* Merged constants with develop.

* Resolved indentation error.

* Added correct call for ENTITY_THRESHOLD constant.

* Merged latent_features/__init__.py with develop.

* All unit tests passing.

* Updated files to be PEP8 compliant.

* Updated files to be PEP8 compliant.

* Updated files to be PEP8 compliant.
