+++ ampligraph/latent_features/models/__init__.py
@@ -1,0 +1,8 @@
+from .EmbeddingModel import EmbeddingModel 
+from .TransE import TransE
+from .DistMult import DistMult
+from .ComplEx import ComplEx
+from .HolE import HolE
+from .RandomBaseline import RandomBaseline
+
+__all__ = ['EmbeddingModel', 'TransE', 'DistMult', 'ComplEx', 'HolE', 'RandomBaseline']
